backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.1
httpx==0.27.2
python-multipart==0.0.12
supabase==2.9.1
google-genai==0.3.0
PyJWT[crypto]==2.9.0
python-dotenv==1.0.1
apscheduler==3.10.4
tzdata==2024.2
